requirements.txt
Flask
pytesseract
Pillow
python-docx
